Change most compilation errors including imports and moving ICalendarCommand into command package
