+++ Calendar/src/calendar/controller/commands/CreateSingleTimedEventCommand.java
@@ -1,21 +1,23 @@
 package calendar.controller.commands;
 
-import model.ICalendarModel;
-import model.SingleEvent;
+import calendar.model.ICalendarModel;
+import calendar.model.IEvent;
+import calendar.model.SingleEvent;
+
 import java.time.LocalDateTime;
 
 public class CreateSingleTimedEventCommand extends CreateEventCommand {
-    public CreateSingleTimedEventCommand(String subject, LocalDateTime startDateTime, LocalDateTime endDateTime) {
-        super(subject, startDateTime, endDateTime, null, null, null);
-    }
+  public CreateSingleTimedEventCommand(String subject, LocalDateTime startDateTime, LocalDateTime endDateTime) {
+    super(subject, startDateTime, endDateTime, null, null, null);
+  }
 
-    @Override
-    public void execute(ICalendarModel calendarModel) {
-        IEvent event = SingleEvent.getBuilder()
-                .subject(subject)
-                .startDateTime(startDateTime)
-                .endDateTime(endDateTime)
-                .build();
-        calendarModel.createEvent(event);
-    }
+  @Override
+  public void execute(ICalendarModel calendarModel) {
+    IEvent event = SingleEvent.getBuilder()
+            .subject(subject)
+            .startDateTime(startDateTime)
+            .endDateTime(endDateTime)
+            .build();
+    calendarModel.createEvent(event);
+  }
 }
